+++ Assets/Scripts/Data/Route.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ad92954293d3a0647af2c1e74d9e14b9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
