+++ Assets/Scripts/Manager.cs
@@ -8,6 +8,7 @@ using System.Collections.ObjectModel;
 using Rails.Rendering;
 using Rails.Controls;
 using Rails.Data;
+using Rails.Systems;
 
 #if UNITY_EDITOR
 using UnityEditor;
